fix: test and move keyringSnapRemovalWarning to its own folder
