+++ ui/components/app/snaps/keyring-snap-removal-warning/index.ts
@@ -1,0 +1,1 @@
+export { default } from './keyring-snap-removal-warning';
