+++ ui/components/app/snaps/snap-removal-result/index.ts
@@ -1,0 +1,1 @@
+export { defatul } from './snap-removal-result';
